feat: handle secret overrides
